feat(processing): refactor patch modules
